(Still broken) PostgreSQL & CockroachDB support

+++ src/main/java/org/springframework/samples/petclinic/model/BaseEntity.java
@@ -15,12 +15,11 @@
  */
 package org.springframework.samples.petclinic.model;
 
-import java.io.Serializable;
-
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.MappedSuperclass;
+import java.io.Serializable;
 
 /**
  * Simple JavaBean domain object with an id property. Used as a base class for objects
@@ -33,13 +32,13 @@ import javax.persistence.MappedSuperclass;
 public class BaseEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
