+++ src/main/java/org/springframework/samples/petclinic/owner/OwnerRepository.java
@@ -15,12 +15,12 @@
  */
 package org.springframework.samples.petclinic.owner;
 
-import java.util.Collection;
-
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Collection;
 
 /**
  * Repository class for <code>Owner</code> domain objects All method names are compliant with Spring Data naming
@@ -51,7 +51,7 @@ public interface OwnerRepository extends Repository<Owner, Integer> {
      */
     @Query("SELECT owner FROM Owner owner left join fetch owner.pets WHERE owner.id =:id")
     @Transactional(readOnly = true)
-    Owner findById(@Param("id") Integer id);
+    Owner findById(@Param("id") Long id);
 
     /**
      * Save an {@link Owner} to the data store, either inserting or updating it.
