+++ src/test/java/org/springframework/samples/petclinic/owner/PetControllerTests.java
@@ -38,8 +38,8 @@ import org.springframework.test.web.servlet.MockMvc;
                             type = FilterType.ASSIGNABLE_TYPE))
 public class PetControllerTests {
 
-    private static final int TEST_OWNER_ID = 1;
-    private static final int TEST_PET_ID = 1;
+    private static final long TEST_OWNER_ID = 1L;
+    private static final long TEST_PET_ID = 1L;
 
 
     @Autowired
@@ -54,7 +54,7 @@ public class PetControllerTests {
     @Before
     public void setup() {
         PetType cat = new PetType();
-        cat.setId(3);
+        cat.setId(3L);
         cat.setName("hamster");
         given(this.pets.findPetTypes()).willReturn(Lists.newArrayList(cat));
         given(this.owners.findById(TEST_OWNER_ID)).willReturn(new Owner());
